refactor: ChessGame의 상태를 관리하는 state와 GameCommand의 패키지 위치를 controller로 변경
