+++ src/test/java/chess/model/game/GameCommandTest.java
@@ -3,6 +3,7 @@ package chess.model.game;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.controller.GameCommand;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
