+++ src/test/java/chess/model/game/state/EndTest.java
@@ -4,8 +4,10 @@ import static chess.helper.PositionFixture.A1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.controller.GameCommand;
+import chess.controller.state.End;
+import chess.controller.state.GameState;
 import chess.model.dto.PlayDto;
-import chess.model.game.GameCommand;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
