+++ src/test/java/chess/model/game/state/PlayTest.java
@@ -6,9 +6,13 @@ import static chess.helper.PositionFixture.A3;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.controller.GameCommand;
+import chess.controller.state.End;
+import chess.controller.state.GameState;
+import chess.controller.state.Play;
+import chess.controller.state.Ready;
 import chess.model.dto.PlayDto;
 import chess.model.game.ChessGame;
-import chess.model.game.GameCommand;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
