+++ src/test/java/chess/model/position/PositionTest.java
@@ -3,8 +3,8 @@ package chess.model.position;
 import static chess.helper.PositionFixture.A1;
 import static chess.helper.PositionFixture.B2;
 import static chess.helper.PositionFixture.D4;
-import static chess.model.position.File.*;
-import static chess.model.position.Rank.*;
+import static chess.model.position.File.A;
+import static chess.model.position.Rank.FIRST;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
