+++ src/test/java/chess/model/position/RankPositionTest.java
@@ -1,8 +1,8 @@
 package chess.model.position;
 
 import static chess.helper.PositionFixture.A1;
-import static chess.model.position.File.*;
-import static chess.model.position.Rank.*;
+import static chess.model.position.File.A;
+import static chess.model.position.Rank.FIRST;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
